// blitzkarte/models/GameData.ts
export interface GameData {
  adjustments_day: string;
  adjustments_span: number;
  adjustments_time: Date;
  administrators: any[];
  assignment_method: string;
  assignments: any[];
  blind_administrators: boolean;
  concurrent_games_limit: number;
  current_year: number;
  deadline_type: string;
  display_as_admin: boolean;
  final_readiness_check: boolean;
  game_name: string;
  game_status: string;
  game_time_zone: number;
  hidden_game: boolean;
  nmr_tolerance_adjustments: boolean | null;
  nmr_tolerance_orders: boolean | null;
  nmr_tolerance_retreats: boolean | null;
  nmr_tolerance_total: number;
  nominations_day: string;
  nominations_span: number;
  nominations_time: Date;
  observe_dst: boolean;
  orders_day: string;
  orders_span: number;
  orders_time: Date;
  partial_roster_start: boolean;
  private_game: boolean;
  retreats_day: string;
  retreats_span: number;
  retreats_time: Date;
  rules: any[];
  start_time: Date;
  stylized_start_year: number;
  time_created: Date;
  turn_1_timing: string;
  vote_delay_count: number | null;
  vote_delay_display_count: number | null;
  vote_delay_display_percent: number | null;
  vote_delay_enabled: boolean;
  vote_delay_lock: number | null;
  vote_delay_percent: number | null;
  votes_day: string;
  votes_span: number;
  votes_time: Date;
  nomination_timing: string;
  nomination_year: number;
  automatic_assignments: boolean;
  rating_limits_enabled: boolean;
  fun_min: number;
  fun_max: number;
  skill_min: number;
  skill_max: number;
};